Move mean_elements to better place and to avoid circular import
